+++ src/gui/model/model.py
@@ -1,0 +1,2 @@
+from gui.view import main_window
+from gui.model import model